+++ .travis.yml
@@ -1,4 +1,4 @@
-# Origin: https://github.com/ah-/rdkafka-dotnet
+# Origin: https://github.com/confluentinc/confluent-kafka-dotnet
 
 language: csharp
 sudo: required
@@ -19,6 +19,6 @@ install:
    - dotnet restore
 
 script:
-   - dotnet build ./src/RdKafka/project.json
-   - dotnet pack ./src/RdKafka/project.json
-   - dotnet test ./test/RdKafka.Tests/
+   - dotnet build ./src/Confluent.Kafka/project.json
+   - dotnet pack ./src/Confluent.Kafka/project.json
+   - dotnet test ./test/Confluent.Kafka.Tests/
