+++ src/Confluent.Kafka/Internal/SafeKafkaHandle.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Runtime.InteropServices;
 
-namespace RdKafka.Internal
+namespace Confluent.Kafka.Internal
 {
     enum RdKafkaType
     {
