+++ src/Confluent.Kafka/TopicConfig.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
-using RdKafka.Internal;
+using Confluent.Kafka.Internal;
 
-namespace RdKafka
+namespace Confluent.Kafka
 {
     /// <summary>
     /// Topic-specific configuration.
@@ -46,7 +46,7 @@ namespace RdKafka
         /// it may be called multiple times for the same message/key.
         ///
         /// Partitioner function constraints:
-        ///   - MUST NOT call any RdKafka methods except for 
+        ///   - MUST NOT call any Confluent.Kafka methods except for
         ///     <see cref="Topic.PartitionAvailable">Topic.PartitionAvailable</see>
         ///   - MUST NOT block or execute for prolonged periods of time.
         ///   - MUST return a value between 0 and partition_cnt-1, or the
@@ -57,7 +57,7 @@ namespace RdKafka
 
         /// <summary>
         /// Sets a custom <see cref="TopicConfig.Partitioner">Partitioner</see>
-        /// delegate to control assignment of messages to partitions. 
+        /// delegate to control assignment of messages to partitions.
         ///
         /// See <see cref="Topic.Produce">Topic.Produce</see> for details.
         /// </summary>
