+++ src/Confluent.Kafka/project.json
@@ -3,8 +3,8 @@
 
     "packOptions": {
         "description": "C# Apache Kafka client",
-        "authors": ["Andreas Heider"],
-        "tags": ["kafka", "rdkafka"]
+        "authors": ["Andreas Heider", "Confluent Inc"],
+        "tags": ["kafka", "librdkafka", "confluent"]
     },
 
     "dependencies": {
@@ -12,7 +12,6 @@
     },
 
     "frameworks": {
-        "net451": { },
         "netstandard1.3": {
             "dependencies": {
                 "System.Console": "4.0.0",
