+++ test/Confluent.Kafka.Tests/project.json
@@ -1,6 +1,6 @@
 {
     "dependencies": {
-        "RdKafka": {
+        "Confluent.Kafka": {
             "target": "project"
         }
     },
@@ -15,10 +15,9 @@
                 "xunit": "2.1.0",
                 "dotnet-test-xunit": "1.0.0-rc2-build10025"
             },
-	    "imports": [
-		    "dnxcore50",
-		    "portable-net45+win8"
-	    ]
+	        "imports": [
+		        "dnxcore50"
+	        ]
         }
     },
     "testRunner": "xunit"
